requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
mysqlclient==2.2.4
packaging==24.1
python-dotenv==1.0.1
Werkzeug==3.0.3
